requirements.txt
crewai
crewai-tools
langchain-openai
python-dotenv
pytube
youtube-transcript-api